requirements.txt
asgiref==3.2.10
Django==3.0.7
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
psycopg2-binary==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
